+++ runtimeconfig/manager_test.go
@@ -2,7 +2,6 @@ package runtimeconfig
 
 import (
 	"context"
-	"crypto/sha256"
 	"fmt"
 	"io"
 	"os"
@@ -76,27 +75,63 @@ func newTestOverridesManagerConfig(t *testing.T, i int32) (*atomic.Int32, Config
 	}
 }
 
+func generateRuntimeFiles(overrideStrings []string) ([]*os.File, error) {
+	var overrideFiles []*os.File
+
+	for count, override := range overrideStrings {
+		pattern := fmt.Sprintf("overrides-file-%d", count)
+		tempFile, err := os.CreateTemp("", pattern)
+		if err != nil {
+			return nil, err
+		}
+		_, err = tempFile.WriteString(override)
+		if err != nil {
+			return nil, err
+		}
+		overrideFiles = append(overrideFiles, tempFile)
+	}
+	return overrideFiles, nil
+}
+
+func generateLoadPath(overrideFiles []*os.File) string {
+	var fileNames []string
+	for _, f := range overrideFiles {
+		fileNames = append(fileNames, f.Name())
+	}
+	return strings.Join(fileNames, ",")
+}
+
+func cleanup(overrideFiles []*os.File) error {
+	for _, f := range overrideFiles {
+		err := f.Close()
+		if err != nil {
+			return err
+		}
+		err = os.Remove(f.Name())
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
 func TestNewOverridesManager(t *testing.T) {
-	tempFile, err := os.CreateTemp("", "test-validation")
+	tempFiles, err := generateRuntimeFiles(
+		[]string{`overrides:
+  user1:
+    limit2: 150`})
 	require.NoError(t, err)
 
 	defer func() {
-		// Clean up
-		require.NoError(t, tempFile.Close())
-		require.NoError(t, os.Remove(tempFile.Name()))
+		require.NoError(t, cleanup(tempFiles))
 	}()
-
-	_, err = tempFile.WriteString(`overrides:
-  user1:
-    limit2: 150`)
-	require.NoError(t, err)
 
 	defaultTestLimits = &TestLimits{Limit1: 100}
 
 	// testing NewRuntimeConfigManager with overrides reload config set
 	overridesManagerConfig := Config{
 		ReloadPeriod: time.Second,
-		LoadPath:     tempFile.Name(),
+		LoadPath:     generateLoadPath(tempFiles),
 		Loader:       testLoadOverrides,
 	}
 
@@ -109,6 +144,116 @@ func TestNewOverridesManager(t *testing.T) {
 
 	// Make sure test limits were loaded.
 	require.NotNil(t, overridesManager.GetConfig())
+}
+
+func TestOverridesManagerMultipleFilesAppend(t *testing.T) {
+	tempFiles, err := generateRuntimeFiles(
+		[]string{`overrides:
+  user1:
+    limit1: 101`,
+			`overrides:
+  user1:
+    limit2: 102`,
+			`overrides:
+  user2:
+    limit1: 103`,
+			`overrides:
+  user2:
+    limit2: 104`})
+	require.NoError(t, err)
+
+	defer func() {
+		require.NoError(t, cleanup(tempFiles))
+	}()
+
+	// testing NewRuntimeConfigManager with overrides reload config set
+	overridesManagerConfig := Config{
+		ReloadPeriod: time.Second,
+		LoadPath:     generateLoadPath(tempFiles),
+		Loader:       testLoadOverrides,
+	}
+
+	overridesManager, err := New(overridesManagerConfig, nil, log.NewNopLogger())
+	require.NoError(t, err)
+	require.NoError(t, services.StartAndAwaitRunning(context.Background(), overridesManager))
+
+	// Cleaning up
+	require.NoError(t, services.StopAndAwaitTerminated(context.Background(), overridesManager))
+
+	// Make sure test limits were loaded.
+	require.NotNil(t, overridesManager.GetConfig())
+	conf := overridesManager.GetConfig().(*testOverrides)
+	require.Equal(t, 101, conf.Overrides["user1"].Limit1)
+	require.Equal(t, 102, conf.Overrides["user1"].Limit2)
+	require.Equal(t, 103, conf.Overrides["user2"].Limit1)
+	require.Equal(t, 104, conf.Overrides["user2"].Limit2)
+}
+
+func TestOverridesManagerMultipleFilesWithOverrides(t *testing.T) {
+	tempFiles, err := generateRuntimeFiles(
+		[]string{
+			`overrides:
+  user1:
+    limit1: 100`,
+			`overrides:
+  user1:
+    limit1: 1234`})
+	require.NoError(t, err)
+
+	defer func() {
+		require.NoError(t, cleanup(tempFiles))
+	}()
+
+	// testing NewRuntimeConfigManager with overrides reload config set
+	overridesManagerConfig := Config{
+		ReloadPeriod: time.Second,
+		LoadPath:     generateLoadPath(tempFiles),
+		Loader:       testLoadOverrides,
+	}
+
+	overridesManager, err := New(overridesManagerConfig, nil, log.NewNopLogger())
+	require.NoError(t, err)
+	require.NoError(t, services.StartAndAwaitRunning(context.Background(), overridesManager))
+
+	// Cleaning up
+	require.NoError(t, services.StopAndAwaitTerminated(context.Background(), overridesManager))
+
+	// Make sure test limits were loaded.
+	require.NotNil(t, overridesManager.GetConfig())
+	conf := overridesManager.GetConfig().(*testOverrides)
+	require.Equal(t, 1234, conf.Overrides["user1"].Limit1)
+}
+
+func TestOverridesManagerMultipleFilesWithEmptyFile(t *testing.T) {
+	tempFiles, err := generateRuntimeFiles(
+		[]string{`overrides:
+  user1:
+    limit1: 100`,
+			``})
+	require.NoError(t, err)
+
+	defer func() {
+		require.NoError(t, cleanup(tempFiles))
+	}()
+
+	// testing NewRuntimeConfigManager with overrides reload config set
+	overridesManagerConfig := Config{
+		ReloadPeriod: time.Second,
+		LoadPath:     generateLoadPath(tempFiles),
+		Loader:       testLoadOverrides,
+	}
+
+	overridesManager, err := New(overridesManagerConfig, nil, log.NewNopLogger())
+	require.NoError(t, err)
+	require.NoError(t, services.StartAndAwaitRunning(context.Background(), overridesManager))
+
+	// Cleaning up
+	require.NoError(t, services.StopAndAwaitTerminated(context.Background(), overridesManager))
+
+	// Make sure test limits were loaded.
+	require.NotNil(t, overridesManager.GetConfig())
+	conf := overridesManager.GetConfig().(*testOverrides)
+	require.Equal(t, 100, conf.Overrides["user1"].Limit1)
 }
 
 func TestManager_ListenerWithDefaultLimits(t *testing.T) {
@@ -143,14 +288,14 @@ func TestManager_ListenerWithDefaultLimits(t *testing.T) {
 	require.NoError(t, services.StartAndAwaitRunning(context.Background(), overridesManager))
 
 	// check if the metrics is set to the config map value before
-	assert.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(fmt.Sprintf(`
+	assert.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
 					# HELP runtime_config_hash Hash of the currently active runtime config file.
 					# TYPE runtime_config_hash gauge
-					runtime_config_hash{sha256="%s"} 1
+					runtime_config_hash{sha256="fa368abb9506cf551faee8f92705fa5d5e1e3cf17ad43bb01f69ff64bac7c301"} 1
 					# HELP runtime_config_last_reload_successful Whether the last runtime-config reload attempt was successful.
 					# TYPE runtime_config_last_reload_successful gauge
 					runtime_config_last_reload_successful 1
-				`, fmt.Sprintf("%x", sha256.Sum256(config))))))
+				`)))
 
 	// need to use buffer, otherwise loadConfig will throw away update
 	ch := overridesManager.CreateListenerChannel(1)
@@ -179,14 +324,14 @@ func TestManager_ListenerWithDefaultLimits(t *testing.T) {
 	require.Equal(t, 100, to.Overrides["user2"].Limit1) // from defaults
 
 	// check if the metrics have been updated
-	assert.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(fmt.Sprintf(`
+	assert.NoError(t, testutil.GatherAndCompare(reg, strings.NewReader(`
 					# HELP runtime_config_hash Hash of the currently active runtime config file.
 					# TYPE runtime_config_hash gauge
-					runtime_config_hash{sha256="%s"} 1
+					runtime_config_hash{sha256="8ff7f344466630e84a3fab44273e8797ec5799dd29f6fdbeea99e4bd771be9bd"} 1
 					# HELP runtime_config_last_reload_successful Whether the last runtime-config reload attempt was successful.
 					# TYPE runtime_config_last_reload_successful gauge
 					runtime_config_last_reload_successful 1
-				`, fmt.Sprintf("%x", sha256.Sum256(config))))))
+				`)))
 
 	// Cleaning up
 	require.NoError(t, services.StopAndAwaitTerminated(context.Background(), overridesManager))